llama3_2/pyproject.toml
[project]
name = "llama3-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.4.2",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "pypdf2>=3.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
